Create Virtual To-Do List
A Virtual To-Do List application that helps users to assemble tasks. It provides a menu with various options. It can add new tasks, track a task and display it's index or add it if not present, display all tasks, and remove tasks on request. It is a command-line application made using Python.